refactor(reposicao): extrai camada de dados de AdminRoutePlanner pra hook (1089→286)
Fatia final do split do RoutePlanner: move todo o state, loaders Supabase,
handlers de check-in/out, geocoding progressivo, allStops/optimizedRoute e
memos de stats pra um hook useRoutePlanner. Movimento puro, behavior-preserving
— lógica copiada verbatim, sem mudança de comportamento.

A página mantém apenas os 4 refs do Leaflet + os 2 efeitos do mapa (init +
markers, acoplados ao DOM, inalterados — leem loading/optimizedRoute do hook)
+ o JSX. Vira um container de ~286 LoC.

Acumulado do RoutePlanner (4 PRs): 1535 → 286 LoC na página (-81%). O hook
(907L) é lógica de dados coesa, não um god-component de UI.

Hook fica fora do strict (herda os `any` pré-existentes da página, igual antes).
